+++ opendrive2lanelet/converter.py
@@ -15,7 +15,7 @@ from opendrive2lanelet.utils import encode_road_section_lane_width_id
 __author__ = "Benjamin Orthen"
 __copyright__ = "TUM Cyber-Physical Systems Group"
 __credits__ = ["Priority Program SPP 1835 Cooperative Interacting Automobiles"]
-__version__ = "1.0.2"
+__version__ = "1.1.0"
 __maintainer__ = "Benjamin Orthen"
 __email__ = "commonroad-i06@in.tum.de"
 __status__ = "Released"
@@ -107,11 +107,13 @@ class OpenDriveConverter:
                 )
 
                 # Create outer lane border
-                outer_parametric_lane_border = OpenDriveConverter._create_outer_lane_border(
-                    lane_borders, lane, coeff_factor
-                )
+                # outer_parametric_lane_border =
 
-                lane_borders.append(outer_parametric_lane_border)
+                lane_borders.append(
+                    OpenDriveConverter._create_outer_lane_border(
+                        lane_borders, lane, coeff_factor
+                    )
+                )
 
                 plane_group = ParametricLaneGroup(
                     id_=encode_road_section_lane_width_id(