+++ opendrive2lanelet/io/osm_convert.py
@@ -7,17 +7,18 @@ import os
 import sys
 import argparse
 
-from pyproj import Proj
 from lxml import etree
 from commonroad.common.file_writer import CommonRoadFileWriter
+from commonroad.common.file_reader import CommonRoadFileReader
 
 from opendrive2lanelet.osm.osm2lanelet import OSM2LConverter
 from opendrive2lanelet.osm.parser import OSMParser
+from opendrive2lanelet.osm.lanelet2osm import L2OSMConverter
 
 __author__ = "Benjamin Orthen"
 __copyright__ = "TUM Cyber-Physical Systems Group"
 __credits__ = ["Priority Program SPP 1835 Cooperative Interacting Automobiles"]
-__version__ = "1.0.2"
+__version__ = "1.1.0"
 __maintainer__ = "Benjamin Orthen"
 __email__ = "commonroad-i06@in.tum.de"
 __status__ = "Released"
@@ -29,7 +30,9 @@ def parse_arguments():
     Returns:
       args object
     """
-    parser = argparse.ArgumentParser(description="Provide ls")
+    parser = argparse.ArgumentParser(
+        description="Parse an OSM file containing lanelets and convert it to a CommonRoad scenario file or vice versa."
+    )
     parser.add_argument("input_file", help="file to be converted")
     parser.add_argument(
         "-f",
@@ -39,6 +42,23 @@ def parse_arguments():
     )
     parser.add_argument("-o", "--output-name", help="specify name of outputed file")
     parser.add_argument("--proj", help="proj-string to specify conversion")
+    parser.add_argument(
+        "-a",
+        "--adjacencies",
+        action="store_true",
+        help="detect left and right adjacencies of lanelets if they do not share a common way",
+    )
+    parser.add_argument(
+        "-r",
+        "--reverse",
+        action="store_true",
+        help="convert in the reverse direction, i.e. from CommonRoad to OSM lanelets.",
+    )
+    parser.add_argument(
+        "--left-driving",
+        action="store_true",
+        help="set to true if map describes a left driving system (e.g. in Great Britain)",
+    )
     args = parser.parse_args()
     return args
 
@@ -48,10 +68,12 @@ def main():
     args = parse_arguments()
     partition = args.input_file.rpartition(".")
 
+    file_ending = "osm" if args.reverse else "xml"
+
     if args.output_name:
         output_name = args.output_name
     else:
-        output_name = f"{partition[0]}.xml"  # only name of file
+        output_name = f"{partition[0]}.{file_ending}"  # only name of file
 
     if os.path.isfile(output_name) and not args.force_overwrite:
         print(
@@ -60,20 +82,63 @@ def main():
         )
         sys.exit(-1)
 
-    with open("{}".format(args.input_file), "r") as file_in:
+    if args.reverse:
+        commonroad_to_osm(args, output_name)
+    else:
+        osm_to_commonroad(args, output_name)
+
+
+def osm_to_commonroad(args, output_name: str):
+    """Convert from OSM to CommonRoad.
+    Args:
+      args: Object containing parsed arguments.
+      output_name: Name of file where result should be written to.
+    """
+    with open(f"{args.input_file}", "r") as file_in:
         parser = OSMParser(etree.parse(file_in).getroot())
         osm = parser.parse()
 
-    osm2l = OSM2LConverter(args.proj)
-    scenario = osm2l(osm)
-    writer = CommonRoadFileWriter(
-        scenario=scenario,
-        planning_problem_set=None,
-        author="",
-        affiliation="",
-        source="OpenDRIVE 2 Lanelet Converter",
-        tags="",
+    osm2l = OSM2LConverter(proj_string=args.proj)
+    scenario = osm2l(
+        osm, detect_adjacencies=args.adjacencies, left_driving_system=args.left_driving
     )
+    if scenario:
+        writer = CommonRoadFileWriter(
+            scenario=scenario,
+            planning_problem_set=None,
+            author="",
+            affiliation="",
+            source="OSM 2 CommonRoad Converter",
+            tags="",
+        )
+        with open(f"{output_name}", "w") as file_out:
+            writer.write_scenario_to_file_io(file_out)
 
-    with open(f"{output_name}", "w") as file_out:
-        writer.write_scenario_to_file_io(file_out)
+    else:
+        print("Could not convert from OSM to CommonRoad format!")
+
+
+def commonroad_to_osm(args, output_name: str):
+    """Convert from CommonRoad to OSM.
+
+    Args:
+      args: Object containing parsed arguments.
+      output_name: Name of file where result should be written to.
+    """
+    try:
+        commonroad_reader = CommonRoadFileReader(args.input_file)
+        scenario, _ = commonroad_reader.open()
+
+    except etree.XMLSyntaxError as xml_error:
+        print(f"SyntaxERror: {xml_error}")
+        print(
+            "There was an error during the loading of the selected CommonRoad file.\n"
+        )
+    l2osm = L2OSMConverter(args.proj)
+    osm = l2osm(scenario)
+    with open(f"{output_name}", "wb") as file_out:
+        file_out.write(
+            etree.tostring(
+                osm, xml_declaration=True, encoding="UTF-8", pretty_print=True
+            )
+        )