+++ opendrive2lanelet/opendriveparser/elements/eulerspiral.py
@@ -7,7 +7,7 @@ from scipy import special
 __author__ = "Benjamin Orthen, Stefan Urban"
 __copyright__ = "TUM Cyber-Physical Systems Group"
 __credits__ = ["Priority Program SPP 1835 Cooperative Interacting Automobiles"]
-__version__ = "1.0.2"
+__version__ = "1.1.0"
 __maintainer__ = "Benjamin Orthen"
 __email__ = "commonroad-i06@in.tum.de"
 __status__ = "Released"
@@ -68,20 +68,25 @@ class EulerSpiral:
 
         else:
             # Fresnel integrals
-            Sa, Ca = special.fresnel(
-                (kappa0 + self._gamma * s) / np.sqrt(np.pi * np.abs(self._gamma))
-            )
-            Sb, Cb = special.fresnel(kappa0 / np.sqrt(np.pi * np.abs(self._gamma)))
-
-            # Euler Spiral
-            Cs1 = np.sqrt(np.pi / np.abs(self._gamma)) * np.exp(
-                1j * (theta0 - kappa0 ** 2 / 2 / self._gamma)
-            )
-            Cs2 = np.sign(self._gamma) * (Ca - Cb) + 1j * Sa - 1j * Sb
-
-            Cs = C0 + Cs1 * Cs2
+            Cs = self._calc_fresnel_integral(s, kappa0, theta0, C0)
 
         # Tangent at each point
         theta = self._gamma * s ** 2 / 2 + kappa0 * s + theta0
 
         return (Cs.real, Cs.imag, theta)
+
+    def _calc_fresnel_integral(self, s, kappa0, theta0, C0):
+        Sa, Ca = special.fresnel(
+            (kappa0 + self._gamma * s) / np.sqrt(np.pi * np.abs(self._gamma))
+        )
+        Sb, Cb = special.fresnel(kappa0 / np.sqrt(np.pi * np.abs(self._gamma)))
+
+        # Euler Spiral
+        Cs1 = np.sqrt(np.pi / np.abs(self._gamma)) * np.exp(
+            1j * (theta0 - kappa0 ** 2 / 2 / self._gamma)
+        )
+        Cs2 = np.sign(self._gamma) * (Ca - Cb) + 1j * Sa - 1j * Sb
+
+        Cs = C0 + Cs1 * Cs2
+
+        return Cs